qcacld-3.0: Handle SSR Reinit failure gracefully
qcacld-3.0 to prima propagation

Presently on the re-init failure when the driver tries
to unregister the wiphy during the re_init failure
SSR timeout happens because the wiphy_unregister
would be waiting on the interface opencount to be set
to 0, it would happen only if the netdevice's are de-registered
with the kernel before wiphy_unregister

De-register the interfaces during the reinit failure so that
upper layer is blocked from accessing the driver and also have
a clean rmmod of the driver.

Change-Id: I3d97df19f6c59fcfd1d423d5cc4eae85040d4a97
CRs-Fixed: 2019989

qcacld-3.0: Apps crash due to invalid wiphy check during unload

Wiphy is not getting unregister at the time of unload because of
that kernel has stale pointer which is causing a crash.

Add a proper check for wiphy during unload and unregister the wiphy.

Change-Id: Ibbdc3966f37bc628552c0118297f82065fa04d46
CRs-Fixed: 2027470